pyproject.toml
[project]
name = "codeleap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "faker>=37.3.0",
    "ipython>=9.2.0",
    "model-bakery>=1.20.4",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "requests>=2.32.3",
    "ruff>=0.11.10",
]
[tool.uv]
cache-dir = "./.uv_cache"
[tool.taskipy.settings]
use_vars = true
[tool.taskipy.tasks]
migrate = "uv run python manage.py migrate"
makemigrations = "uv run python manage.py makemigrations"
runserver = "uv run python manage.py runserver 8004"
createsuperuser = "uv run python manage.py createsuperuser"
shell_plus = "uv run python manage.py shell_plus"
showmigrations = "uv run python manage.py showmigrations"
pytest = "uv run pytest -vv"
lint = "ruff check"
lint_fix = "ruff check --fix"
[tool.ruff]
exclude = [
  # No need to traverse our git directory
  ".git",
  ".gitignore",
  ".gitattributes",
  ".gitmodules",
  ".gitkeep",
  ".gitlab-ci.yml", # There's no value in checking cache directories
  "__pycache__", # The conf file is mostly autogenerated, ignore it
  "docs/source/conf.py", # The old directory contains Flake8 2.0
  "old", # This contains our built documentation
  "build", # This contains builds of flake8 that we don't want to check
  "dist", # Ignore venv
  "venv",
  ".venv", # Ignore migrations
  "migrations", # Ignore media
  "media", # Ignore static
  "static", # Ignore templates
  "templates",
  "src",
  "config",
  "scripts", # Ignore temporary files
  "*.swp",
  "tmp",
  "temp",
  ".DS_Store",
  ".vscode",
  ".idea",
]
line-length = 120
[tool.ruff.lint]
ignore = ["PT009", "F403", "F405"]
select = ["E4", "E7", "E9", "F"]
[tool.flake8]
exclude = [
  # No need to traverse our git directory
  ".git",
  ".gitignore",
  ".gitattributes",
  ".gitmodules",
  ".gitkeep",
  ".gitlab-ci.yml", # There's no value in checking cache directories
  "__pycache__", # The conf file is mostly autogenerated, ignore it
  "docs/source/conf.py", # The old directory contains Flake8 2.0
  "old", # This contains our built documentation
  "build", # This contains builds of flake8 that we don't want to check
  "dist", # Ignore venv
  "venv",
  ".venv", # Ignore migrations
  "migrations", # Ignore media
  "media", # Ignore static
  "static", # Ignore templates
  "templates",
  "src",
  "config",
  "scripts", # Ignore temporary files
  "*.swp",
  "tmp",
  "temp",
  ".DS_Store",
  ".vscode",
  ".idea",
]
extend-ignore = ["F401", "F403", "F405"]
max-complexity = 13
max-line-length = 120
